+++ spec/mongoid/errors/invalid_field_spec.rb
@@ -4,19 +4,27 @@ describe Mongoid::Errors::InvalidField do
 
   describe "#message" do
 
+    module MyModule
+      def crazy_method; self; end
+    end
+
+    before do
+      Object.send(:include, MyModule)
+    end
+
     let(:error) do
-      described_class.new("collection")
+      described_class.new(Person, :crazy_method)
     end
 
     it "contains the problem in the message" do
       error.message.should include(
-        "Defining a field named collection is not allowed."
+        "Defining a field named 'crazy_method' is not allowed."
       )
     end
 
     it "contains the summary in the message" do
       error.message.should include(
-        "Defining fields that conflict with Mongoid internal attributes"
+        "Defining this field would override the method 'crazy_method'"
       )
     end